+++ eslint-config-dsx-base/prettier.js
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: [
+    './rules/prettier',
+  ].map(require.resolve),
+  rules: {},
+};