+++ eslint-config-dsx-base/CHANGELOG.md
@@ -1,3 +1,9 @@
+2.0.0 / 2020-04-13
+==================
+
+- [breaking] add typescript-eslint
+- [breaking] split prettier
+
 1.0.0 / 2020-04-13
 ==================